Allow namespaced data types
puppet-lint detects namespaced data types as CLASSREF instead of TYPE
because lexically they are identical. Since the built in data types are
hardcoded and detected as TYPE, we can assume that if the CLASSREF
includes :: it is probably a namespaced data type and thus valid syntax.
This means unnamespaced custom data types are unsupported, but that's
reserved for Puppet built-ins anyway.

This also drops the s.type != :TYPE check because it's redundant.
s.type can't be both (:NAME or :CLASSREF) and :TYPE.